Add WagmiProvider + QueryClientProvider
