+++ package.json
@@ -50,7 +50,9 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@safe-global/safe-kit": "file:../safe-core-sdk/packages/safe-kit"
+    "@safe-global/safe-kit": "file:../safe-core-sdk/packages/safe-kit",
+    "viem": "^2.18.6",
+    "wagmi": "^2.12.2"
   },
   "packageManager": "yarn@1.22.22"
 }
