+++ src/hooks/useConfig.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 
-import { SafeContext } from '../context.js'
+import { SafeContext } from '../context.jsx'
 import type { ConfigParam, SafeConfig } from '../types/index.js'
 
 export type UseConfigParams<Config extends SafeConfig = SafeConfig> = ConfigParam<Config>
