+++ src/index.ts
@@ -1,3 +1,3 @@
-export { SafeProvider, SafeContext } from './context.js'
+export { SafeProvider, SafeContext } from './context.jsx'
 export * from './hooks/index.js'
 export * from './types/index.js'
