+++ tsconfig.build.json
@@ -3,6 +3,7 @@
   "include": ["src/**/*.ts"],
   "exclude": ["src/**/*.test.ts", "src/**/*.test-d.ts"],
   "compilerOptions": {
-    "sourceMap": true
+    "sourceMap": true,
+    "jsx": "preserve"
   }
 }
